refactor: deduplicate poster notification logic in poster routes (#52)
Extract shared YouTube URL parsing and presenter notification logic from
poster and poster-bigpicture routes into a reusable posterShowNotification
helper, reusing existing buildPosterNotification from presenterNotifications.
